Fix multiple compile errors and warnings on MSVC
There are still plenty of warnings with default flags, but these should
be harmless, nothing to spend time on now.

Fixes #8.

+++ bwidgets/generics/bwPolygon.h
@@ -27,7 +27,7 @@ class bwPolygon {
 
  protected:
   bwPointVec vertices{};
-  unsigned long int vert_count{0};
+  size_t vert_count{0};
 };
 
 }  // namespace bWidgets
